edit transaction done (have not reviewed freq transactions)

+++ application/models/VO/Transaction.php
@@ -5,7 +5,7 @@ class VO_Transaction extends Zend_Db_Table_Row {
             $other = Proxy_Transaction::getInstance()->retrieveAllByFreqTranId($this->id_freq_tran);
             if($other->count() == 1){//remove freq_tran if there is no child left
                 $freq = Proxy_FreqTran::getInstance()->findById($this->id_freq_tran);
-                $freq->delete();
+                if($freq) $freq->delete();//it could be removed when the user edited it
             }
         }
         return parent::delete();
